pg_txn_dashboard/requirements.txt
streamlit>=1.44
psycopg[binary]>=3.1
pandas>=2.0
python-dotenv>=1.0
